General cleanup - Added cpp's for editor code - Moved SceneSystem to the Engine - Fixed Quat euler angles issue in ctools
