+++ Editor/source/panels/ScenePanel.h
@@ -1,71 +1,16 @@
-#ifndef ScenePanel_h__
-#define ScenePanel_h__
+#pragma once
 
 #include "flEngine.h"
 #include "EditorSystem.h"
-#include "SceneSystem.h"
 #include "ctVector.h"
 
 class ScenePanel : public flEngine::GUI::Panel
 {
-  class GUIVisitor : public flEngine::Scene::Visitor<flEngine::Scene::Node>
-  {
-  public:
-    GUIVisitor(int64_t selectedID) : m_selectedID(selectedID) {}
-
-    virtual bool OnEnter(flIN flEngine::Scene::Node* pNode) {
-      if (pNode == pNode->GetScene()->GetRootNode())
-        return true;
-
-      flEngine::GUI::Widgets::PushID(pNode->GetID());
-
-      bool open = flEngine::GUI::Widgets::BeginTreeNode(pNode->GetName(), m_selectedID == pNode->GetID());
-      if (flEngine::GUI::Widgets::IsItemClicked()) {
-        m_selectedID = m_selectedID == pNode->GetID() ? -1 : pNode->GetID();
-      }
-
-      if (!open)
-        flEngine::GUI::Widgets::PopID();
-
-      return open;
-    }
-
-    virtual void OnLeave(flIN flEngine::Scene::Node* pNode) {
-      if (pNode != pNode->GetScene()->GetRootNode())
-      {
-        flEngine::GUI::Widgets::EndTreeNode();
-        flEngine::GUI::Widgets::PopID();
-      }
-    }
-
-    int64_t m_selectedID;
-  };
-
 public:
-  ScenePanel(flEngine::GUI::GUISystem* pGUI)
-    : Panel(pGUI, "Scene")
-  {
-    m_pSceneSystem = flEngine::Application::Get().GetSubSystem<SceneSystem>();;
-  }
+  ScenePanel(flEngine::GUI::GUISystem* pGUI);
 
-  virtual void OnGUI() override
-  {
-    flEngine::Ref<EditorSystem> pEditor = flEngine::Application::Get().GetSubSystem<EditorSystem>();
-    flEngine::Ref<flEngine::Scene::SceneGraph> pScene = m_pSceneSystem->ActiveScene();
-    if (flEngine::GUI::Widgets::Button("Add")) {
-      flEngine::Scene::Node* pNode = pScene->AddNode("New Node", pEditor->m_selectedNode);
-      pEditor->m_selectedNode = pNode->GetID();
-    }
-
-    flEngine::GUI::Widgets::Separator();
-
-    GUIVisitor visitor(pEditor->m_selectedNode);
-    pScene->Traverse(&visitor, nullptr);
-    pEditor->m_selectedNode = visitor.m_selectedID;
-  }
+  virtual void OnGUI() override;
 
 private:
-  flEngine::Ref<SceneSystem> m_pSceneSystem;
+  flEngine::SceneSystem *m_pSceneSystem = nullptr;
 };
-
-#endif // ScenePanel_h__
